@@ -1,5 +1,5 @@
 {
-    "name":        "hampe/zurb-ink-bundle",
+    "name":        "stafox/zurb-ink-bundle",
     "type":        "symfony-bundle",
     "description": "Creating email templates is hard. This Bundle provides help.",
     "keywords":    ["email", "zurb", "ink", "template"],
@@ -10,20 +10,25 @@
             "name": "Thomas Hampe",
             "homepage": "http://github.com/thampe",
             "email": "github@hampe.co"
+        },
+        {
+            "name": "Stas Kuryan",
+            "homepage": "http://github.com/Stafox",
+            "email": "stas.kuryan@gmail.com"
         }
     ],
     "require": {
-        "php": "^5.3.2 || ^7.0",
+        "php": "^7.0",
         "symfony/framework-bundle": "^2.0 || ^3.0",
-        "twig/twig": "^1.23 || ^2.0",
-        "tijsverkoyen/css-to-inline-styles": "^1.0",
+        "twig/twig": "^2.0",
+        "voku/css-to-inline-styles": "2.0.0",
         "hampe/inky": "^1.3.6"
     },
     "autoload": {
         "psr-0": {
-            "Hampe\\Bundle\\ZurbInkBundle": ""
+            "Stafox\\ZurbInkBundle": ""
         }
     },
 
-    "target-dir": "Hampe/Bundle/ZurbInkBundle"
+    "target-dir": "Stafox/ZurbInkBundle"
 }
